requirements.txt
faiss-cpu
sentence-transformers
transformers
gradio
pypdf
numpy